Update main.py to use the docs folder, instead of public. Remove /public/ from the repo. Update .gitignore to ignore the docs folder

+++ src/main.py
@@ -1,25 +1,41 @@
 import os
 import shutil
+import sys
 
 from markdownblock import markdown_to_html_node, extract_title
 
 
 def main():
-    copy_static_to_public()
-    generate_pages_recursive("content/", "template.html", "public/")
 
+    base_path = "./"
 
-def copy_static_to_public():
-    public_path = "./public/"
-    static_path = "./static/"
+    # Get base path
+    if len(sys.argv) > 1:
+        base_path = sys.argv[1]
 
+    # Create paths
+    public_path = os.path.join(base_path, "docs/")
+
+    static_path = os.path.join(base_path, "static/")
+    content_path = os.path.join(base_path, "content/")
+    template_path = os.path.join(base_path, "template.html")
+
+    # Check paths exist
     if not os.path.exists(static_path):
-        raise Exception("Error: The static folder doesn't exist")
+        raise Exception(f'The static folder doesn\'t exist. "{static_path}"')
+    if not os.path.exists(content_path):
+        raise Exception(f'The content folder doesn\'t exist. "{content_path}"')
+    if not os.path.exists(template_path):
+        raise Exception(f'The HTML template is missing. "{template_path}"')
 
-    shutil.rmtree(public_path)
+    # Clear the public folder
+    if os.path.exists(public_path):
+        shutil.rmtree(public_path)
     os.mkdir(public_path)
 
+    # Populate public folder
     copy_dir_to_dir(static_path, public_path)
+    generate_pages_recursive(base_path, content_path, template_path, public_path)
 
 
 def copy_dir_to_dir(source: str, destination: str):
@@ -35,7 +51,22 @@ def copy_dir_to_dir(source: str, destination: str):
             copy_dir_to_dir(src_joined_path, tgt_joined_path)
 
 
-def generate_page(from_path: str, template_path: str, dest_path: str):
+def generate_pages_recursive(
+    base_path: str, dir_path_content: str, template_path: str, dest_dir_path: str
+):
+
+    for item in os.listdir(dir_path_content):
+
+        path = os.path.join(dir_path_content, item)
+        dest_path = os.path.join(dest_dir_path, item)
+
+        if os.path.isdir(path):
+            generate_pages_recursive(base_path, path, template_path, dest_path)
+        elif os.path.isfile(path) and ".md" in path:
+            generate_page(base_path, path, template_path, dest_path[:-3] + ".html")
+
+
+def generate_page(base_path: str, from_path: str, template_path: str, dest_path: str):
     print(f"Generating page from {from_path} to {dest_path} using {template_path}")
 
     if from_path is None or template_path is None or dest_path is None:
@@ -54,6 +85,8 @@ def generate_page(from_path: str, template_path: str, dest_path: str):
 
     template_contents = template_contents.replace("{{ Title }}", title)
     template_contents = template_contents.replace("{{ Content }}", html_node.to_html())
+    template_contents = template_contents.replace('href="/', f'href="{base_path}')
+    template_contents = template_contents.replace('src="/', f'src="{base_path}')
 
     dest_path_split = dest_path.split("/")
     if not os.path.isfile(dest_path_split[0]):
@@ -70,21 +103,6 @@ def generate_page(from_path: str, template_path: str, dest_path: str):
     with open(dest_path, "w+") as f:
         f.write(template_contents)
         f.close()
-
-
-def generate_pages_recursive(
-    dir_path_content: str, template_path: str, dest_dir_path: str
-):
-
-    for item in os.listdir(dir_path_content):
-
-        path = os.path.join(dir_path_content, item)
-        dest_path = os.path.join(dest_dir_path, item)
-
-        if os.path.isdir(path):
-            generate_pages_recursive(path, template_path, dest_path)
-        elif os.path.isfile(path) and ".md" in path:
-            generate_page(path, template_path, dest_path[:-3] + ".html")
 
 
 main()
